improved some testing
